♻️ move to src-layout and scikit-build-core
