--- MANIFEST.in
@@ -1,21 +1,0 @@
-prune docs
-prune test
-
-graft extern/mqt-core
-prune **/.github
-prune **/apps
-prune **/doc
-prune **/docs
-prune **/meta
-prune **/plots
-prune **/test
-prune **/tests
-
-prune extern/mqt-core/extern/json/include
-prune extern/mqt-core/extern/googletest
-prune extern/mqt-core/extern/boost/config/checks
-prune extern/mqt-core/extern/boost/config/tools
-prune extern/mqt-core/extern/boost/multiprecision/config
-prune extern/mqt-core/extern/boost/multiprecision/example
-prune extern/mqt-core/extern/boost/multiprecision/performance
-prune extern/mqt-core/extern/boost/multiprecision/tools
