+++ pyproject.toml
@@ -1,11 +1,6 @@
 [build-system]
-requires = [
-    "setuptools>=61",
-    "setuptools_scm[toml]>=7",
-    "ninja>=1.10; sys_platform != 'win32'",
-    "cmake>=3.19",
-]
-build-backend = "setuptools.build_meta"
+requires = ["scikit-build-core>=0.5.0", "setuptools-scm>=7", "pybind11>=2.11"]
+build-backend = "scikit_build_core.build"
 
 [project]
 name = "mqt.qecc"
@@ -79,37 +74,65 @@ Documentation = "https://github.com/cda-tum/mqt-qecc"
 "Bug Tracker" = "https://github.com/cda-tum/mqt-qecc/issues"
 Discussions = "https://github.com/cda-tum/mqt-qecc/discussions"
 
-[tool.setuptools.packages.find]
-include = ["mqt.*"]
+[tool.scikit-build]
+# Protect the configuration against future changes in scikit-build-core
+minimum-version = "0.5.0"
 
-[tool.setuptools_scm]
+# Set the target to build
+cmake.targets = ["pyqecc"]
 
-[tool.cibuildwheel]
-build = "cp3*"
-archs = "auto64"
-skip = "*-musllinux*"
-test-skip = "*-macosx_arm64 cp312-*"
-test-command = "python -c \"from mqt import qecc\""
-environment = { DEPLOY = "ON" }
-build-frontend = "build"
-build-verbosity = 3
-manylinux-x86_64-image = "manylinux_2_28"
+# Set required CMake and Ninja versions
+cmake.minimum-version = "3.19"
+ninja.minimum-version = "1.10"
 
-[tool.cibuildwheel.linux]
-before-all = [
-    "yum -y install mpfr-devel",
-    "git clone --branch v2.9.0 --depth 1 https://github.com/flintlib/flint2.git",
-    "cd flint2 && ./configure && make -j 3 && make install"
+# Setuptools-style build caching in a local directory
+build-dir = "build/{wheel_tag}"
+
+# Build stable ABI wheels for CPython 3.12+
+wheel.py-api = "cp312"
+
+# Explicitly set the package directory
+wheel.packages = ["src/mqt"]
+
+metadata.version.provider = "scikit_build_core.metadata.setuptools_scm"
+sdist.include = ["src/mqt/qecc/_version.py"]
+sdist.exclude = [
+    "**/.github",
+    "**/doc",
+    "**/docs",
+    "**/meta",
+    "**/plots",
+    "**/test",
+    "**/tests",
+    "extern/mqt-core/extern/json/include",
+    "extern/mqt-core/extern/googletest",
+    "extern/mqt-core/extern/boost/config/checks",
+    "extern/mqt-core/extern/boost/config/tools",
+    "extern/mqt-core/extern/boost/multiprecision/config",
+    "extern/mqt-core/extern/boost/multiprecision/example",
+    "extern/mqt-core/extern/boost/multiprecision/performance",
+    "extern/mqt-core/extern/boost/multiprecision/tools"
 ]
 
-[tool.cibuildwheel.macos]
-archs = "x86_64"
-environment = { MACOSX_DEPLOYMENT_TARGET = "10.15", DEPLOY = "ON" }
+[tool.scikit-build.cmake.define]
+BUILD_MQT_QECC_TESTS = "OFF"
+BUILD_MQT_QECC_BINDINGS = "ON"
+ENABLE_IPO = "ON"
 
-#[tool.cibuildwheel.windows] TODO
-#before-build = "pip install delvewheel"
-#repair-wheel-command = "delvewheel repair -v -w {dest_dir} {wheel}"
-#environment = { DEPLOY = "ON", CC = "clang", CXX = "clang++", FLINT_ROOT = "/clang64" }
+
+[tool.check-sdist]
+sdist-only = ["src/mqt/qecc/_version.py"]
+git-only = [
+    "docs/*",
+    "extern/*",
+    "examples/*",
+    "test/*",
+    ".idea/*",
+]
+
+
+[tool.setuptools_scm]
+write_to = "src/mqt/qecc/_version.py"
 
 [tool.pytest.ini_options]
 minversion = "7.2"
@@ -147,7 +170,8 @@ profile = "black"
 src_paths = ["mqt/qecc", "test/python"]
 
 [tool.mypy]
-files = ["mqt/qecc", "test/python", "setup.py"]
+files = ["src/mqt", "test/python"]
+mypy_path = ["$MYPY_CONFIG_FILE_DIR/src"]
 python_version = "3.8"
 strict = true
 warn_unused_configs = true
