+++ scripts/numerics/code_construct/constructHGPcodes.py
@@ -1,4 +1,6 @@
 """Hypergraph product code construction from Roffe's LDPC library."""
+from __future__ import annotations
+
 import ldpc.protograph as pt
 import numpy as np
 from bposd.hgp import hgp
