+++ scripts/numerics/code_construct/constructToricCodes.py
@@ -1,5 +1,7 @@
 """Toric code construction."""
 # using tools from https://github.com/quantumgizmos/bp_osd
+from __future__ import annotations
+
 import numpy as np
 from bposd.hgp import hgp
 from ldpc.codes import ring_code
