+++ src/mqt/qecc/cc_decoder/comparison/tn_decoder.py
@@ -1,4 +1,6 @@
 """Decoding simulation using the tensor network implementation of the qecsim package."""
+from __future__ import annotations
+
 import json
 from pathlib import Path
