+++ src/mqt/qecc/cc_decoder/run.sh
