+++ src/mqt/qecc/pyqecc.pyi
