+++ src/python/bindings.cpp
@@ -25,9 +25,6 @@
 
 #include <pybind11/stl.h>
 
-#define STRINGIFY(x) #x
-#define MACRO_STRINGIFY(x) STRINGIFY(x)
-
 namespace py = pybind11;
 using namespace pybind11::literals;
 
@@ -195,10 +192,4 @@ PYBIND11_MODULE(pyqecc, m) {
           "circuit_name"_a,
           "ecc_name"_a,
           "ecc_frequency"_a = 100);
-
-#ifdef VERSION_INFO
-    m.attr("__version__") = MACRO_STRINGIFY(VERSION_INFO);
-#else
-    m.attr("__version__") = "dev";
-#endif
 }
